More accuracy in the fragment shader output.
Apparently, it is not required to support highp in the fragment shader.  So switched to using mediump.  This fix should support more devices at the higher precission while taking the consequence of less accuracy over the 32 bit per channel surface.

+++ app/src/main/cpp/levelDrawer/common.hpp
@@ -50,7 +50,7 @@ namespace renderDetails {
         ~ParametersNormalMap() override = default;
     };
 
-    using PostprocessingDataInputGL = boost::variant<std::vector<uint32_t>, std::vector<uint8_t>>;
+    using PostprocessingDataInputGL = boost::variant<std::vector<uint16_t>, std::vector<uint8_t>>;
 }
 
 namespace levelDrawer {
